+++ src/main/java/org/bigml/binding/utils/StemmerInterface.java
@@ -1,0 +1,5 @@
+package org.bigml.binding.utils;
+
+public interface StemmerInterface {
+    public String getStem(String term);
+}